chore: move format to its own module
